+++ engine/src/main/java/net/jqwik/engine/facades/ArbitrariesFacadeImpl.java
@@ -252,28 +252,9 @@ public class ArbitrariesFacadeImpl extends Arbitraries.ArbitrariesFacade {
 		Function<Arbitrary<T>, Arbitrary<T>> recur,
 		int depth
 	) {
-		if (depth <= 0) {
+		if (depth == 0) {
 			return base.get();
 		}
-		final Arbitrary<T> inner = recursive(base, recur, depth - 1);
-		final Arbitrary<T> outer = recur.apply(inner);
-
-		return new Arbitrary<T>() {
-			@Override
-			public RandomGenerator<T> generator(int genSize) {
-				return outer.generator(genSize);
-			}
-
-			@Override
-			public EdgeCases<T> edgeCases() {
-				// Due to combinatorial explosion
-				// TODO: remove as soon as edge cases are created on the fly
-				// see https://github.com/jlink/jqwik/issues/114
-				if (depth > 5) {
-					return EdgeCases.none();
-				}
-				return outer.edgeCases();
-			}
-		};
+		return recur.apply(recursive(base, recur, depth - 1));
 	}
 }